feat: extract saves functions
