+++ bin/v1.4.3/static
@@ -1,0 +1,1 @@
+../../static/
