+++ configure.ac
@@ -1,5 +1,5 @@
 # Define the package name and version
-AC_INIT([helapordo], [1.4.2], [jgabaut@github.com])
+AC_INIT([helapordo], [1.4.3-dev], [jgabaut@github.com])
 
 # Verify automake version and enable foreign option
 AM_INIT_AUTOMAKE([foreign -Wall])
@@ -169,7 +169,7 @@ AM_CONDITIONAL([LINUX_BUILD], [test "$build_linux" = "yes"])
 # Set a default version number if not specified externally
 AC_ARG_VAR([VERSION], [Version number])
 if test -z "$VERSION"; then
-  VERSION="1.4.2"
+  VERSION="1.4.3-dev"
 fi
 
 # Output variables to the config.h header
