+++ src/consumables.h
@@ -1,0 +1,8 @@
+#ifndef CONSUMABLES_H_
+#define CONSUMABLES_H_
+
+#include "game_core.h"
+
+void initConsumableBag(Fighter * f, Koliseo * kls);
+
+#endif // CONSUMABLES_H_
