+++ src/equips.h
@@ -1,0 +1,8 @@
+#ifndef EQUIPS_H_
+#define EQUIPS_H_
+
+#include "game_core.h"
+
+void applyEquipPerks(Equip * e, Fighter * f);
+
+#endif // EQUIPS_H_
