+++ src/specials.c
@@ -904,6 +904,40 @@ void assassinSpecial_Venomblade(WINDOW *w, Fighter *f, Enemy *e, Boss *b,
         wattroff(w, COLOR_PAIR(S4C_BRIGHT_YELLOW));
     }
 }
+
+/**
+ * Takes a Fighter pointer and prepares its specialSlot fields by allocating SPECIALSMAX slots.
+ * Special slots are initialised.
+ * The specific move assigned is determined using the ordering of both fighterClass values and SPECIALSMAX.
+ * @see Fighter
+ * @see Specialslot
+ * @see SPECIALSMAX
+ * @see REGISTER_CALLBACK()
+ * @see costFromSpecial()
+ * @see stringFromSpecial()
+ * @param kls The Koliseo used for allocations.
+ * @param f The Fighter pointer whose special slots will be initialised.
+ */
+void setSpecials(Fighter *f, Koliseo *kls)
+{
+    char movename[80];
+    char movedesc[80];
+    for (int i = 0; i <= SPECIALSMAX; i++) {
+        kls_log(kls, "DEBUG", "Prepping Specialslot (%i)", i);
+        Specialslot *s =
+            (Specialslot *) KLS_PUSH_TYPED(kls, Specialslot, HR_Specialslot,
+                                           "Specialslot", "Specialslot");
+        s->enabled = 0;
+        s->move = i + (f->class * (SPECIALSMAX + 1));	// Assign the i-th move offsetting by classNum * specialsMax
+        s->cost = costFromSpecial(f->class, i);
+        strcpy(movename, nameStringFromSpecial(f->class, i));
+        strcpy(movedesc, descStringFromSpecial(f->class, i));
+        //printf("DEBUG\n%i\t%s\n",(i+ (f->class * (SPECIALSMAX + 1))),stringFromSpecial(f->class,i));
+        strcpy(s->name, movename);
+        strcpy(s->desc, movedesc);
+        f->specials[i] = s;
+    };
+}
 #else
 #ifndef HELAPORDO_RAYLIB_BUILD
 #error "HELAPORDO_CURSES_BUILD and HELAPORDO_RAYLIB_BUILD are both undefined.\n"
@@ -999,6 +1033,10 @@ void assassinSpecial_Disguise(Rectangle * win, Fighter * f, Enemy * e, Boss * b,
 void assassinSpecial_Venomblade(Rectangle * win, Fighter * f, Enemy * e, Boss * b,
                                 Path * p, int roomIndex, int enemyIndex,
                                 int isBoss)
+{
+
+}
+void setSpecials(Fighter *f, Koliseo *kls)
 {
 
 }
