+++ stego.lock
@@ -62,3 +62,4 @@ errortestsdir = "kulpo"
 "1.4.0" = "Add WIP raylib integration, bump s4c to 0.4.2"
 "1.4.1" = "Fix broken clang build, add s4c-scripts, bump s4c to 0.4.3, invil to 0.2.5"
 "1.4.2" = "Change static_path to working dir, Gamestate has screen for curses build, check screen size, bump deps"
+"1.4.3" = "Foo"
